ai-service/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
PyPDF2
python-docx
pytest
